requirements.txt
# Core ML / LLM
openai>=1.12.0
transformers>=4.37.0
torch>=2.1.0
accelerate>=0.26.0
bitsandbytes>=0.42.0
sentencepiece>=0.1.99
protobuf>=4.25.0

# Data processing
pandas>=2.1.0
numpy>=1.26.0

# Evaluation metrics
scikit-learn>=1.4.0
pingouin>=0.5.4
scipy>=1.12.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
tabulate>=0.9.0

# Visualization and reporting
matplotlib>=3.8.0
seaborn>=0.13.0
